+++ src/Raven.CodeAnalysis/Binder/BlockBinder.MemberAccess.cs
@@ -1231,7 +1231,7 @@ partial class BlockBinder
             var substituted = SubstituteTypeParameters(nullableType.UnderlyingType, substitutions);
             return SymbolEqualityComparer.Default.Equals(substituted, nullableType.UnderlyingType)
                 ? type
-                : substituted.GetNullableType();
+                : substituted.ApplySubstitutedNullability(nullableType);
         }
 
         if (type is RefTypeSymbol refTypeType)